PENDING: fix(ti): AM62L Failed to get response error messages
Since AM62L doesn't implement the get_device in TIFS, the TI SCI message
in bl31_setup is bound to return errors.
This may cause users to get alarmed as they may see errors in initial
boot logs.
Move the ti_sci_get_revision and ti_sci_device_get in
soc specific folders.
